+++ vue/src/components/CenterArea/MainContents/Schedule/ScheduledTweetsList.vue
@@ -1,9 +1,11 @@
 <script setup>
 import { ref, onBeforeMount } from 'vue'
 import { useCurrentUserStore } from '@/stores/currentUser.js'
+import { useRouter } from 'vue-router'
 import axios from 'axios'
 import ScheduledTweet from './ScheduledTweet.vue'
 
+const router = useRouter()
 const currentUser = useCurrentUserStore()
 const accounts = ref(null)
 const tweets = ref(null)
@@ -28,13 +30,18 @@ async function fetchData() {
   }
 }
 async function deleteTweet(id) {
-  try {
-    await axios.delete('http://localhost:8081/api/schedule/' + id)
-    window.location.reload()
-  } catch (err) {
-    error.value = err.response ? `${err.response.status}: ${err.response.statusText}` : err.message
-  } finally {
-    fetchData()
+  const confirmed = window.confirm('この予約ツイートを削除しますか？')
+  if (confirmed) {
+    try {
+      await axios.delete('http://localhost:8081/api/schedule/' + id)
+      router.replace({ name: 'schedule', params: { userId: currentUser.userId } }).then(() => {
+        fetchData()
+      })
+    } catch (err) {
+      error.value = err.response
+        ? `${err.response.status}: ${err.response.statusText}`
+        : err.message
+    }
   }
 }
 onBeforeMount(() => {
